pyproject.toml
[project]
name = "repl-nix-workspace"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.11"
dependencies = [
    "email-validator>=2.2.0",
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "openai>=1.58.1",
    "psycopg2-binary>=2.9.10",
    "pypdf2>=3.0.1",
    "python-docx>=1.1.2",
    "sqlalchemy>=2.0.36",
    "trafilatura>=2.0.0",
    "werkzeug>=3.1.3",
]
